Move skip_numba_jit() and skip_sdc_jit() to test_utils.py
